REFACTOR new folder structure for klant features
