--- src/features/brp/index.ts
@@ -1,1 +1,0 @@
-export { default as PersoonZoeker } from "./PersoonZoeker.vue";
