--- src/features/brp/types.ts
@@ -1,8 +1,0 @@
-export interface Persoon {
-  voornamen: string;
-  voorvoegsel: string;
-  achternaam: string;
-  leeftijd: number | null;
-  id: string;
-  bsn: string;
-}
