--- src/features/klant/config.ts
@@ -1,1 +1,0 @@
-export const KLANT_SELECTED = "KlantSelected";
