+++ src/features/klant/index.ts
@@ -1,4 +1,5 @@
-export * from "./config";
 export * from "./service";
-export { default as KlantDetails } from "./KlantDetails.vue";
-export { default as KlantZoeker } from "./KlantZoeker.vue";
+export { default as PersoonDetails } from "./persoon/PersoonDetails.vue";
+export { default as PersoonZoeker } from "./persoon/PersoonZoeker.vue";
+export { default as BrpGegevens } from "./persoon/BrpGegevens.vue";
+export { usePersoonByBsn } from "./persoon/service";
