+++ src/features/klant/types.ts
@@ -1,5 +1,3 @@
-import type { ServiceData } from "@/services";
-
 export type UpdateContactgegevensParams = Pick<
   Klant,
   "id" | "telefoonnummers" | "emails"
@@ -17,34 +15,7 @@ export interface Klant {
   bsn?: string;
 }
 
-export interface Persoon {
-  _typeOfKlant: "persoon";
-  bsn: string;
-  postcode?: string;
-  huisnummer?: string;
-  geboortedatum?: Date;
-  voornaam: string;
-  voorvoegselAchternaam?: string;
-  achternaam: string;
-  geboorteplaats: string;
-  geboorteland: string;
-}
-
 export enum KlantType {
   Persoon = "natuurlijk_persoon",
   Bedrijf = "vestiging",
-}
-
-export interface EnrichedPersoon {
-  naam: ServiceData<string | null>;
-  bsn: string | undefined;
-  telefoonnummers: ServiceData<string | null>;
-  emails: ServiceData<string | null>;
-  geboortedatum: ServiceData<Date | null | undefined>;
-  postcodeHuisnummer: ServiceData<string | null>;
-  create: () => Promise<void>;
-  detailLink: ServiceData<{
-    to: string;
-    title: string;
-  } | null>;
 }
