+++ src/views/PersoonDetailView.vue
@@ -11,7 +11,7 @@
       </ul>
     </nav>
     <simple-spinner v-if="klant.loading" />
-    <klant-details v-else-if="klant.success" :klant="klant.data" />
+    <persoon-details v-else-if="klant.success" :klant="klant.data" />
     <application-message
       v-else
       message="Er is geen klant gevonden"
@@ -19,7 +19,7 @@
     ></application-message>
 
     <simple-spinner v-if="klantBsn && persoon.loading" />
-    <klant-brp-gegevens
+    <brp-gegevens
       v-if="persoon.success && persoon.data"
       :persoon="persoon.data"
     />
@@ -84,16 +84,19 @@ import {
   ContactmomentenOverzicht,
   useContactverzoekenByKlantId,
 } from "@/features/contactmoment";
-import { KlantDetails, useKlantById } from "@/features/klant";
+import {
+  PersoonDetails,
+  useKlantById,
+  BrpGegevens,
+  usePersoonByBsn,
+} from "@/features/klant";
 import ApplicationMessage from "@/components/ApplicationMessage.vue";
 import SimpleSpinner from "@/components/SimpleSpinner.vue";
-import ContactverzoekenOverzicht from "../features/contactmoment/ContactverzoekenOverzicht.vue";
-import Pagination from "../nl-design-system/components/Pagination.vue";
+import ContactverzoekenOverzicht from "@/features/contactmoment/ContactverzoekenOverzicht.vue";
+import Pagination from "@/nl-design-system/components/Pagination.vue";
 import { useContactmomentenByKlantId } from "@/features/shared/get-contactmomenten-service";
 import { useZakenByBsn } from "@/features/zaaksysteem";
-import ZakenOverzicht from "../features/zaaksysteem/ZakenOverzicht.vue";
-import KlantBrpGegevens from "../features/klant/brp/KlantBrpGegevens.vue";
-import { usePersoonByBsn } from "@/features/klant/brp/service";
+import ZakenOverzicht from "@/features/zaaksysteem/ZakenOverzicht.vue";
 
 const props = defineProps<{ klantId: string }>();
 const klantId = computed(() => props.klantId);
